agro/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
django==5.1.1
django-cpf-cnpj==1.0.0
django-filter==24.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.4
markdown==3.7
markupsafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
pip==24.2
psycopg2-binary==2.9.9
pytz==2024.2
pyyaml==6.0.2
requests==2.32.3
simplejson==3.19.3
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
uv==0.4.12
